grand def/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pyserial==3.5
Werkzeug==2.0.3
pytest==7.4.0
pytest-mock==3.11.1